[Wallet] Fix a tx time mismatch after an orphan
There was a tenacious bug where a transaction would take the time of an
older orphaned transaction in the wallet.
After analysis, the new transaction seems to be spending the same UTXO
than the orphaned transaction. So in substance the wallet value for the
time of the transaction was kept from the first attempt at spending the
UTXO.
We fix the problem by resetting the time value of the transaction in the
wallet to 0 when it's orphaned.